requirements.txt
Flask==1.1.2
pylint
pandas
google-cloud-language
wikipedia